requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
brotli==1.2.0
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
dj-database-url==3.1.2
Django==6.0
drf_spectacular
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
gunicorn==25.1.0
idna==3.11
iniconfig==2.3.0
kombu==5.6.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.12.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
